backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pgvector==0.3.5
alembic==1.14.0
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.3.0
redis[asyncio]==5.2.0
anthropic==0.40.0
sentence-transformers==3.3.1
python-multipart==0.0.12
python-dotenv==1.0.1
apscheduler==3.10.4
pydantic==2.10.0
pydantic-settings==2.6.0
pypdf==5.1.0
python-docx==1.1.2
tenacity==9.0.0
rich==13.9.0
email-validator==2.2.0
